[资源] 资源下载功能（包含已下载判断、积分扣减、资源作者积分增加）

+++ xinder-blog-api/src/main/java/com/xinder/api/bean/Resources.java
@@ -40,7 +40,7 @@ public class Resources implements Serializable {
     private String url;
 
     @ApiModelProperty(value = "上传用户")
-    private Integer uid;
+    private Long uid;
 
     @ApiModelProperty(value = "下载所需积分")
     private Integer point;
